+++ R/ridge2.R
@@ -20,15 +20,18 @@
 #' "blockbootstrap", "movingblockbootstrap", "splitconformal" (very experimental right now),
 #' "rvinecopula" (with Gaussian margins for now, Student-t coming soon)
 #' @param block_length Length of block for circular or moving block bootstrap
-#' @param margins Distribution of margins "student" (postponed) or "gaussian" for
-#' \code{type_pi == "rvinecopula"}
+#' @param margins Distribution of margins: "gaussian", "empirical", "student" (postponed or
+#' never) for \code{type_pi == "rvinecopula"}
 #' @param seed Reproducibility seed for random stuff
 #' @param B Number of bootstrap replications or number of simulations (yes, 'B' is unfortunate)
 #' @param type_aggregation Type of aggregation, ONLY for bootstrapping; either "mean" or "median"
 #' @param centers Number of clusters for \code{type_clustering}
 #' @param type_clustering "kmeans" (K-Means clustering) or "hclust" (Hierarchical clustering)
+#' @param ym Univariate time series (\code{stats::ts}) of yield to maturities with
+#' \code{frequency = frequency(y)} and \code{start = tsp(y)[2] + 1 / frequency(y)}.
+#' Default is \code{NULL}.
 #' @param cl An integer; the number of clusters for parallel execution, for bootstrap
-#' @param ... Additional parameters to be passed \code{\link{kmeans}} or \code{\link{hclust}}
+#' @param ... Additional parameters to be passed to \code{\link{kmeans}} or \code{\link{hclust}}
 #'
 #' @return An object of class "mtsforecast"; a list containing the following elements:
 #'
@@ -133,12 +136,13 @@ ridge2f <- function(y,
                       "splitconformal"
                     ),
                     block_length = NULL,
-                    margins = "gaussian", # or "student"
+                    margins = c("gaussian", "empirical", "student"),
                     seed = 1,
                     B = 100L,
                     type_aggregation = c("mean", "median"),
                     centers = NULL,
                     type_clustering = c("kmeans", "hclust"),
+                    ym = NULL,
                     cl = 1L,
                     ...)
 {
@@ -474,6 +478,16 @@ ridge2f <- function(y,
       }
     }
 
+    if (!is.null(ym))
+    {
+      neutralized_sims <- lapply(series_names,
+                                 function(series)
+                                   neutralize(out, ym = ym,
+                                              selected_series = series))
+      names(neutralized_sims) <- series_names
+      out$neutralized_sims <- neutralized_sims
+    }
+
     return(structure(out, class = "mtsforecast"))
   }
 
@@ -522,6 +536,8 @@ ridge2f <- function(y,
 
   if (identical(type_pi, "rvinecopula"))
   {
+    margins <- match.arg(margins)
+
     cl <- floor(min(max(cl, 0L), parallel::detectCores()))
 
     if (cl <= 1L)
@@ -675,6 +691,16 @@ ridge2f <- function(y,
       }
     }
 
+    if (!is.null(ym))
+    {
+      neutralized_sims <- lapply(series_names,
+                                 function(series)
+                                   neutralize(out, ym = ym,
+                                              selected_series = series))
+      names(neutralized_sims) <- series_names
+      out$neutralized_sims <- neutralized_sims
+    }
+
     return(structure(out, class = "mtsforecast"))
   }
 
@@ -703,7 +729,9 @@ fit_ridge2_mts <- function(x,
                              "rvinecopula",
                              "splitconformal"
                            ),
-                           margins = "gaussian", # or "student"
+                           margins = c("gaussian",
+                                       "empirical",
+                                       "student"),
                            ...)
 {
   stopifnot(floor(nb_hidden) == nb_hidden)
@@ -713,10 +741,12 @@ fit_ridge2_mts <- function(x,
   nodes_sim <- match.arg(nodes_sim)
   activ <- match.arg(activ)
   type_pi <- match.arg(type_pi)
-  stopifnot(margins %in% c("gaussian", "student"))
-  margins <- switch(match.arg(margins), # this is convoluted
+  stopifnot(margins %in% c("gaussian", "empirical", "student"))
+  choice_margins <- match.arg(margins)
+  margins <- switch(choice_margins, # this is convoluted
                     gaussian = "normal",
-                    student = "t")
+                    student = "t",
+                    empirical = "empirical")
 
   series_names <- colnames(x)
 
@@ -823,7 +853,7 @@ fcast_ridge2_mts <- function(fit_obj,
                              type_forecast = c("recursive", "direct"),
                              level = 95,
                              type_simulation = c("none", "rvinecopula"), # other than bootstrap
-                             margins = c("gaussian", "student"),
+                             margins = c("gaussian", "empirical", "student"),
                              bootstrap = FALSE,
                              type_bootstrap = c("bootstrap",
                                                 "blockbootstrap",
@@ -835,7 +865,8 @@ fcast_ridge2_mts <- function(fit_obj,
   type_simulation <- match.arg(type_simulation)
   margins <- switch(match.arg(margins),
                    gaussian = "normal",
-                   student = "t")
+                   student = "t",
+                   empirical = "empirical")
 
   if (identical(bootstrap, FALSE))
   {